Version 1.0.5.
In this version the code dump in the Wordpress.org subversion
repository have an empty directory called "_notes". Git can not save
empty directories because it does not track the path structure in that
way. It will be removed and never come back in version 1.0.6 and
becasue of this I think it's rather a mistake and because of this ok
to leave it out.